+++ CHANGELOG.md
@@ -125,6 +125,7 @@
 - Add mypy for static type checking for most of the codebase
 - Now supports latest deluge-web-client (=>2.x.x)
 - Remove support for PTPIMG
+- Optimized UNIT3D torrent re-download with regression testing
 
 ### Fixed